requirements.txt
openai>=0.27.0
selenium>=4.9.0
webdriver-manager>=3.8.6
python-dotenv>=1.0.0
PyQt5>=5.15.9
PyQt5-Qt5>=5.15.2
PyQt5-sip>=12.12.1
QDarkStyle>=3.1.0 